third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/navigating-across-documents/about-srcdoc-navigation-blocked-expected.txt

This is a testharness.js-based test.
FAIL Navigations to about:srcdoc via window.location are blocked. assert_equals: expected "blocked" but got "allowed"
FAIL Navigations to about:srcdoc via window.open() are blocked. assert_equals: expected "blocked" but got "allowed"
Harness: the test ran to completion.
